libathemecore/memory: sstrndup: use size_t not int for length argument
This breaks the ABI on 64-bit but we're in the middle of a development
cycle anyway; master should not be considered stable.

+++ include/atheme_memory.h
@@ -13,7 +13,7 @@ E void *smalloc(size_t size);
 E void *scalloc(size_t elsize, size_t els);
 E void *srealloc(void *oldptr, size_t newsize);
 E char *sstrdup(const char *s);
-E char *sstrndup(const char *s, int len);
+E char *sstrndup(const char *s, size_t len);
 
 #endif
